requirements.txt
streamlit>=1.34
langchain>=0.2
langchain-community>=0.2
langchain-groq>=0.1
groq>=0.9
python-dotenv>=1.0